Add some basic debugging features, fix shell escaping
Shell escaping was broken in the original version, since I misread
the API (it returns a new string instead of substituting in place).

You can turn debugging on by changing Aqualung.scrobbler.debug to
true, and then it will print out the command lines it issues, as
well as an explanation for why it won't scrobble a particular
track.